Creacion validator para deleteProperty

+++ src/controllers/propetyController.js
@@ -63,6 +63,14 @@ export const updateProperty = async (req, res) => {
 // TODO Fix Bug When Trying To Delete A Property That Doesn't Exist
 export const deleteProperty = async (req, res) => {
     try {
+        const { errors } = validationResult(req);
+        // Check If Request Have Some Error
+        if (errors.length != 0) {
+            return res.status(400).json({
+                msg: errors[0].msg
+            });
+        }
+
         const { id } = req.params;
 
         await Property.destroy({
